refactor: [FE] 게임 시작버튼 컴포넌트와 버튼 클릭 관련 기능을 GameStartButton로 분리
- Intent
GameModeList 컴포넌트가 점점 복잡해진다고 생각했고, 게임시작버튼과 관련 로직을 별도로 분리하는게 더 좋아보여서 분리하였습니다.
